Allow decimal quantities in Record Order + show item count
Quantity inputs (search-add and per-line-item) in Record Order used
parseInt and integer-only step/min, silently rounding down anything
like 1.5kg for loose/bulk items. Switched to parseFloat with
step="0.01" so decimal quantities are accepted and calculate
correctly through to billed/cost totals.

Also added an "Items in Order" line to the live summary showing both
line-item count and total units (e.g. "1 item · 3.5 units").

Fixed a few downstream spots that would have silently truncated a
decimal quantity back to a whole number once recorded: the bill's
item rows (admin.js/agent.js, print view and WhatsApp PDF) and the
Shop Insights / top-items analytics in both Admin and Agent Hub, all
switched from parseInt to parseFloat on Quantity.

+++ app/views/admin.js
@@ -644,7 +644,7 @@
                 const itemName = prod ? prod.name : sku;
 
                 if (!itemStats[itemName]) itemStats[itemName] = { totalQty: 0, orderAppearances: 0, lastOrderIndex: orderIndex };
-                itemStats[itemName].totalQty += (parseInt(i['Quantity']) || 1);
+                itemStats[itemName].totalQty += (parseFloat(i['Quantity']) || 1);
                 itemStats[itemName].orderAppearances += 1;
                 itemStats[itemName].lastOrderIndex = orderIndex;
             });
@@ -1881,7 +1881,7 @@
             const sku = (it['SKU'] || '').trim();
             const prod = productMapBySKU[sku];
             const name = prod ? prod.name : (sku || 'Item');
-            const qty = parseInt(it['Quantity']) || 0;
+            const qty = parseFloat(it['Quantity']) || 0;
             const unitPrice = parseFloat(String(it['Unit Price'] || 0).replace(/[^0-9.-]+/g, '')) || 0;
             const lineTotal = parseFloat(String(it['Calculated Total'] || (qty * unitPrice)).replace(/[^0-9.-]+/g, '')) || (qty * unitPrice);
             itemsSubtotal += lineTotal;
@@ -2301,7 +2301,7 @@
             const oc=oidCust[oid]; if(!oc) return;
             const sku=(i['SKU']||'').trim();
             const std = skuToStd[sku] || (productMapBySKU[sku]?productMapBySKU[sku].name:sku);
-            const qty = parseInt(i['Quantity'])||0;
+            const qty = parseFloat(i['Quantity'])||0;
             const up = toNum(i['Unit Price']), ap = toNum(i['Actual Price']);
             const rev = toNum(i['Calculated Total']);
             const mp = up>0 ? ((up-ap)/up*100) : 0;
